Move conic gradient to new file
